requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
streamlit==1.39.0
scikit-learn==1.5.2
google-generativeai==0.8.6
python-dotenv==1.2.2
PyPDF2==3.0.1
python-docx==1.1.2
python-multipart==0.0.12
